+++ changelog.md
@@ -46,6 +46,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0)
 - Fixed keyrepeat not working in tty on linux (#913)
 - Fixed `multi-tap` not holding (#958, #976)
 - Fixed `multi-tap` cancelling on release of other keys (#974)
+- Fixed the `c_src/mac/list-keyboards.c` on Apple SDK < 12.0 (#987)
 
 ## 0.4.3 – 2024-09-11